Move all routes to individual files
