+++ packages/rest/src/RouteImports.ts
@@ -1,0 +1,3 @@
+export * as REST from 'discord-api-types/rest';
+export { Routes } from 'discord-api-types/rest';
+export { group, route, type } from './route-group';
