+++ packages/rest/src/routes/ApplicationCommand.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/* Routes on https://discord.com/developers/docs/interactions/application-commands */
+export const applicationCommand = group({});
