+++ packages/rest/src/routes/Emoji.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/emoji. */
+export const emoji = group({});
