+++ packages/rest/src/routes/Guild.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/guild. */
+export const guild = group({});
