+++ packages/rest/src/routes/GuildScheduledEvent.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/guild-scheduled-event. */
+export const guildScheduledEvent = group({});
