+++ packages/rest/src/routes/GuildTemplate.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/guild-template. */
+export const guildTemplate = group({});
