+++ packages/rest/src/routes/InteractionResponse.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/interactions/receiving-and-responding. */
+export const interactionResponse = group({});
