+++ packages/rest/src/routes/Invite.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/invite. */
+export const invite = group({});
