+++ packages/rest/src/routes/OAuth2.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/topics/oauth2. */
+export const oauth2 = group({});
