+++ packages/rest/src/routes/StageInstance.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/stage-instance. */
+export const stageInstance = group({});
