+++ packages/rest/src/routes/Sticker.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/sticker. */
+export const sticker = group({});
