+++ packages/rest/src/routes/Webhook.ts
@@ -1,0 +1,4 @@
+import '../RouteImports';
+
+/** Routes on https://discord.com/developers/docs/resources/webhook. */
+export const webhook = group({});
