+++ packages/rest/src/routes/index.ts
@@ -1,0 +1,17 @@
+export * from './routes/ApplicationCommand';
+export * from './routes/AuditLog';
+export * from './routes/AutoModeration';
+export * from './routes/Channel';
+export * from './routes/Emoji';
+export * from './routes/Gateway';
+export * from './routes/Guild';
+export * from './routes/GuildScheduledEvent';
+export * from './routes/GuildTemplate';
+export * from './routes/InteractionResponse';
+export * from './routes/Invite';
+export * from './routes/OAuth2';
+export * from './routes/StageInstance';
+export * from './routes/Sticker';
+export * from './routes/User';
+export * from './routes/Voice';
+export * from './routes/Webhook';
